Added custom hook
