+++ src/components/article-params-form/hooks/useClickOutside.tsx
@@ -1,0 +1,22 @@
+import { useEffect } from 'react';
+
+const useClickOutside = (
+	ref: React.RefObject<HTMLElement>,
+	handler: () => void
+) => {
+	useEffect(() => {
+		const handleClickOutside = (event: MouseEvent) => {
+			if (ref.current && !ref.current.contains(event.target as Node)) {
+				handler();
+			}
+		};
+
+		document.addEventListener('mousedown', handleClickOutside);
+
+		return () => {
+			document.removeEventListener('mousedown', handleClickOutside);
+		};
+	}, [ref, handler]);
+};
+
+export default useClickOutside;
